list cleanup
