+++ src/components/LeftColumn/dummyData.js
@@ -1,0 +1,32 @@
+export const initialList = [
+  {
+    id: 1,
+    title: 'How to Start Creating your Brand ...',
+    author: 'Jacob Ruiz',
+    lastEdited: 'Just now',
+  },
+  {
+    id: 2,
+    title: 'Your brand is more than grammar...',
+    author: 'Jacob Ruiz',
+    lastEdited: 'Just now',
+  },
+  {
+    id: 3,
+    title: '10 Great Movies at the New York Fe...',
+    author: 'Jacob Ruiz',
+    lastEdited: 'Just now',
+  },
+  {
+    id: 4,
+    title: 'Brand and Values Exploration',
+    author: 'Jacob Ruiz',
+    lastEdited: 'Just now',
+  },
+  {
+    id: 5,
+    title: 'Demo Doc',
+    author: 'Jacob Ruiz',
+    lastEdited: 'Just now',
+  },
+];
